relocate the testcase
